Add files via upload
A Website where an algorithm will allow us to search the best courses uploaded on youtube. The goal is to filter videos so we can access the best free courses ever uploaded on youtube. The goal of this website is to bypass the tideous task of researching and sorting the best content when it comes to learning and avoid the long search across irrelevant formations, tutorials and courses to select the videos with the best value.